pyproject.toml
[project]
name = "PlaywrightCapture"
version = "1.28.6"
description = "A simple library to capture websites using playwright"
authors = [
    {name="Raphaël Vinot", email= "raphael.vinot@circl.lu"}
]
license = "BSD-3-Clause"
readme = "README.md"
requires-python = ">=3.9"

dynamic = [ "classifiers" ]

dependencies = [
    "playwright (>=1.51.0)",
    "dateparser (>=1.2.1)",
    "beautifulsoup4[charset-normalizer,lxml] (>=4.13.4)",
    "w3lib (>=2.3.1)",
    "tzdata (>=2025.2)",
    "playwright-stealth (>=1.0.6)",
    "setuptools (>=78.1.0)",
    "puremagic (>=1.28)",
    "async-timeout (>=5.0.1) ; python_version < \"3.11\"",
    "aiohttp[speedups] (>=3.11.16)",
    "aiohttp-socks (>=0.10.1)",
    "typing-extensions (>=4.12.2,<5.0.0) ; python_version < \"3.12\"",
    "dnspython (>=2.7.0,<3.0.0)",
    "python-socks (>=2.7.1,<3.0.0)"
]

[project.urls]
Repository = "https://github.com/Lookyloo/PlaywrightCapture"
Issues = "https://github.com/Lookyloo/PlaywrightCapture/issues"

[tool.poetry]
classifiers=[
    'Intended Audience :: Science/Research',
    'Intended Audience :: Telecommunications Industry',
    'Topic :: Security',
    'Topic :: Internet',
]


[project.optional-dependencies]
recaptcha = [
    "pydub (>=0.25.1)",
    "SpeechRecognition (>=3.14.2)"
]

[tool.poetry.group.dev.dependencies]
types-beautifulsoup4 = "^4.12.0.20250204"
pytest = "^8.3.5"
mypy = "^1.15.0"
types-dateparser = "^1.2.0.20250408"
types-pytz = "^2025.2.0.20250326"


[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"
